requirements.txt
numpy
pickleshare
Pillow
python-dateutil
pytorch-fid
pytorch-lightning
scikit-image
scikit-video
scipy
tensorboard
tensorboard-plugin-wit
torch-ema==0.3
torch-fidelity
torchmetrics
tqdm
traitlets
matplotlib
